Backend: Refactor Auth classes
